% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxwr.R
\name{eqxwr_add}
\alias{eqxwr_add}
\title{eqxwr_add}
\usage{
eqxwr_add(df, country = NULL, description = NULL)
}
\arguments{
\item{df}{A data.frame or file name pointing to csv file. The contents of the data.frame or csv file should be exactly two olumns: state, containing a list of all 3125 EQ-5D-5L health state vectors, and a column of corresponding utility values, with a suitable name.}

\item{country}{Optional string. If not NULL, will be used as a country description for the user-defined value set.}

\item{description}{Optional string. If not NULL, will be used as a descriptive text for the user-defined value set.}
}
\value{
True/False, indicating success or error.
}
\description{
Add user-defined EQ-5D-5L value set to reverse crosswalk options.
}
\examples{
# make nonsense value set
new_df <- data.frame(state = make_all_EQ_indexes(), TEST = runif(3125))
# Add as value set for Fantasia
eqxwr_add(new_df, 'Fantasia')
}
