% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_functions.R
\name{eqxwr_drop}
\alias{eqxwr_drop}
\title{eqxwr_drop}
\usage{
eqxwr_drop(country = NULL)
}
\arguments{
\item{country}{Optional string. If NULL, a list of current user-defined value sets will be provided for selection. If set, and matching an existing user-defined value set, a prompt will be given as to whether the value set should be deleted.}
}
\value{
True/False, indicating success or error.
}
\description{
Drop user-defined EQ-5D-5L value set to reverse crosswalk options.
}
\examples{
# make nonsense value set
new_df <- data.frame(state = make_all_EQ_indexes(), TEST = runif(3125))
# Add as value set for Fantasia
eqxwr_add(new_df, 'Fantasia')
# Drop value set for Fantasia
eqxwr_drop('Fantasia')
}
